+++ ASM2_API/ASM-Java5/src/main/java/com/dto/cart/CartCreateRequest.java
@@ -1,0 +1,8 @@
+package com.dto.cart;
+
+public record CartCreateRequest(
+        Long userId,
+        Long productId,
+        Integer quantity
+) {
+}
