+++ ASM2_API/ASM-Java5/src/main/java/com/dto/cart/CartResponse.java
@@ -1,0 +1,12 @@
+package com.dto.cart;
+
+import java.time.OffsetDateTime;
+
+public record CartResponse(
+        Long cartId,
+        Long userId,
+        Long productId,
+        Integer quantity,
+        OffsetDateTime dateAdded
+) {
+}
