+++ ASM2_API/ASM-Java5/src/main/java/com/dto/cart/CartUpdateRequest.java
@@ -1,0 +1,7 @@
+package com.dto.cart;
+
+public record CartUpdateRequest(
+        Long userId,
+        Integer quantity
+) {
+}
