+++ ASM2_API/ASM-Java5/src/main/java/com/dto/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package com.product;
+package com.dto.product;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
