+++ ASM2_API/ASM-Java5/src/main/java/com/dto/product/ProductUpdateRequest.java
@@ -1,11 +1,10 @@
-package com.product;
+package com.dto.product;
 
 import java.math.BigDecimal;
 
-public record ProductCreateRequest(
+public record ProductUpdateRequest(
+        Long productId,
         String productName,
-        Long parentId,
-        Long sellerId,
         Integer categoryId,
         Integer stockCount,
         String thumbnailExtension,
